Select object will auto fill selected column's alias by its field name.
This feature is important to those situations that the NameMeta might be changed in different projects but the field name in class remain unchanged. So that, down streams select from the auto-filled select object can also remain unchanged to refer the items with alias auto-filled.
This feature can be turn off by override Provider.isAutoFillSelectTableColumnAliasByFieldName() returns false.